+++ src/pages/mypage/index.js
@@ -256,6 +256,6 @@ const Span = styled.span`
 `;
 const BottomDiv = styled.div`
   width: 100%;
-  height: 727px;
+  /* height: 727px; */
   background-color: #f7f8f9;
 `;
